Add unit tests and implementation for Kimi K2.7-code thinking and metadata handling
- Introduced tests for fallback model metadata to ensure correct behavior for Kimi K2.7-code, including temperature handling and context/output limits.
- Added tests for thinking payload construction to enforce the requirement that Kimi K2.7-code always uses enabled thinking.
- Implemented logic in `thinking.ts` to manage thinking settings and payloads for various model families, ensuring compliance with API constraints.
- Enhanced schema generation for Kimi K2.7-code to reflect the always-on thinking requirement in the user interface.
